LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/d-expected.txt


FAIL d responsive to inherited d changes assert_equals: expected (string) "path(\"M 0 0 H 300\")" but got (undefined) undefined
